debt - move contrib only service to contrib (output)
