+++ src/vs/workbench/workbench.desktop.main.ts
@@ -66,7 +66,6 @@ import 'vs/workbench/electron-browser/desktop.main';
 
 import 'vs/workbench/services/integrity/node/integrityService';
 import 'vs/workbench/services/search/electron-browser/searchService';
-import 'vs/workbench/services/output/electron-browser/outputChannelModelService';
 import 'vs/workbench/services/textfile/electron-browser/nativeTextFileService';
 import 'vs/workbench/services/keybinding/electron-browser/nativeKeymapService';
 import 'vs/workbench/services/extensions/electron-browser/extensionService';
@@ -118,6 +117,9 @@ registerSingleton(ITunnelService, TunnelService);
 
 
 //#region --- workbench contributions
+
+// Output
+import 'vs/workbench/contrib/output/electron-browser/outputChannelModelService';
 
 // Tags
 import 'vs/workbench/contrib/tags/electron-browser/workspaceTagsService';
